Move `freeze` util from skipruntime_init to skipruntime_utils
